Move helpers to slasHelpers
